requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
bcrypt==4.2.0
black==24.8.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
ciscoconfparse==1.9.51
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
cryptography==43.0.0
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.6.1
dotmap==1.3.30
exceptiongroup==1.2.2
executing==2.0.1
future==1.0.0
hier-config==2.2.3
idna==3.7
importlib_metadata==8.2.0
iniconfig==2.0.0
ipaddr==2.2.0
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
junos-eznc==2.7.1
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy-extensions==1.0.0
napalm==5.0.0
ncclient==0.6.15
netaddr==1.3.0
netmiko==4.4.0
netutils==1.9.1
nornir==3.4.1
nornir-jinja2==0.2.0
nornir-netmiko==1.0.1
nornir-rich==0.1.6
nornir-utils==0.2.0
ntc_templates==6.0.0
orionsdk==0.4.0
packaging==24.1
paramiko==3.4.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyaml==24.7.0
pycparser==2.22
pyeapi==1.0.2
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyserial==3.5
pytest==8.3.2
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scp==0.15.0
six==1.16.0
stack-data==0.6.3
style==1.1.6
tenacity==9.0.0
textfsm==1.1.3
toml==0.10.2
tomli==2.0.1
traitlets==5.14.3
transitions==0.9.2
ttp==0.9.5
ttp-templates==0.3.6
typing_extensions==4.12.2
update==0.0.1
urllib3==2.2.2
wcwidth==0.2.13
wrapt==1.16.0
yamlordereddictloader==0.4.2
zipp==3.20.0
